Widen frames8 to int in ObjSeq_run to drop spurious byte mask
frames8 (s8) feeding a u8* store produced an extra clrlwi truncation that
consumed r4, shifting the six loop-invariant pointer setups up one register
(r5-r10 instead of r4-r9). Declaring it int removes the mask and realigns
the pointer allocation; instruction count now matches target. 94.1->94.8%.

(cherry picked from commit dc4d7f0d19d0d373704bbf695df525ddcf58fc05)

+++ src/main/objseq.c
@@ -457,7 +457,7 @@ void ObjSeq_run(void)
     f32* dists;
     f32* frames;
     u8* marks;
-    s8 frames8;
+    int frames8;
     u8* matched[0x28];
     s16 keepBuf[0x5a];
     int objectCount;
